refactor: AppConfig
